Make grid_1d into a module
This merge also breaks out functions related to `tanh_dz` into
their own module.
